+++ components/navbar/StatsDropdown.vue
